types

+++ _old/Bookstore/resources/item.py
@@ -1,5 +1,5 @@
 from flask_restful import Resource, reqparse
-from flask_jwt_extended import jwt_required, fresh_jwt_required
+from flask_jwt_extended import jwt_required, jwt_optional, get_jwt_claims, get_jwt_identity, fresh_jwt_required
 from models.item import ItemModel
 
 
@@ -16,14 +16,15 @@ class Item(Resource):
                         help="Every item needs a store_id."
                         )
 
-    def get(self, name: str):
+    @jwt_required
+    def get(self, name):
         item = ItemModel.find_by_name(name)
         if item:
             return item.json()
         return {'message': 'Item not found'}, 404
 
     @fresh_jwt_required
-    def post(self, name: str):
+    def post(self, name):
         if ItemModel.find_by_name(name):
             return {'message': "An item with name '{}' already exists.".format(name)}, 400
 
@@ -38,16 +39,20 @@ class Item(Resource):
 
         return item.json(), 201
 
-    @jwt_required
-    def delete(self, name: str):
 
+    @jwt_required
+    def delete(self, name):
+        """Make sure user is admin"""
+        claims = get_jwt_claims()
+        if not claims['is_admin']:
+            return {'message': 'Admin privilege required'}, 401
         item = ItemModel.find_by_name(name)
         if item:
             item.delete_from_db()
             return {'message': 'Item deleted.'}
         return {'message': 'Item not found.'}, 404
 
-    def put(self, name: str):
+    def put(self, name):
         data = Item.parser.parse_args()
 
         item = ItemModel.find_by_name(name)
@@ -63,5 +68,12 @@ class Item(Resource):
 
 
 class ItemList(Resource):
+    @jwt_optional
     def get(self):
-        return {'items': [item.json() for item in ItemModel.find_all()]}, 200
+        """Specify what data is being passed if user is not logged in"""
+        user_id = get_jwt_identity()
+        items = [item.json() for item in ItemModel.find_all()]
+        if user_id:
+            return {'items':items}, 200
+        return {'items': [item['name'] for item in items],
+                'message': 'Price data available if logged in.'}, 200
